chore: refactor to use factory
