+++ packages/create-amplify/src/package-manager-controller/index.ts
@@ -1,4 +1,6 @@
-export { NpmPackageManagerController } from './npm_package_manager_controller.js';
-export { PnpmPackageManagerController } from './pnpm_package_manager_controller.js';
-export { YarnClassicPackageManagerController } from './yarn_classic_package_manager_controller.js';
-export { YarnModernPackageManagerController } from './yarn_modern_package_manager_controller.js';
+export { packageManagerControllerFactory } from './package_manager_controller_factory.js';
+export {
+  PackageManager,
+  PackageManagerName,
+  getPackageManagerFactory,
+} from './package_manager.js';
