--- packages/create-amplify/src/package-manager-controller/package_manager_controller.ts
@@ -1,8 +1,0 @@
-export type DependencyType = 'dev' | 'prod';
-
-export type PackageManagerController = {
-  installDependencies: (
-    packageNames: string[],
-    type: DependencyType
-  ) => Promise<void>;
-};
