+++ packages/create-amplify/src/package-manager-controller/pnpm_package_manager_controller.ts
@@ -1,10 +1,10 @@
 import { execa as _execa } from 'execa';
 import { executeWithDebugLogger } from '../execute_with_logger.js';
-import { type PackageManager } from '../package_manager.js';
+import { type PackageManager } from './package_manager.js';
 import {
   DependencyType,
   PackageManagerController,
-} from './package_manager_controller.js';
+} from './package_manager_controller_factory.js';
 
 /**
  *
