+++ packages/create-amplify/src/project-initializer/index.ts
@@ -1,4 +1,1 @@
-export { NpmProjectInitializer } from './npm_project_initializer.js';
-export { PnpmProjectInitializer } from './pnpm_project_initializer.js';
-export { YarnClassicProjectInitializer } from './yarn_classic_project_initializer.js';
-export { YarnModernProjectInitializer } from './yarn_modern_project_initializer.js';
+export { projectInitializerFactory } from './project_initializer_factory.js';
