+++ packages/create-amplify/src/project-initializer/npm_project_initializer.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, mock } from 'node:test';
 import { NpmProjectInitializer } from './npm_project_initializer.js';
 import assert from 'assert';
-import { packageManagers } from '../package_manager.js';
+import { getPackageManagerFactory } from '../package-manager-controller/index.js';
 
 void describe('InitializedEnsurer', () => {
   void it('does nothing if package.json already exists', async () => {
@@ -9,7 +9,7 @@ void describe('InitializedEnsurer', () => {
     const execaMock = mock.fn();
     const npmInitializedEnsurer = new NpmProjectInitializer(
       '/testProjectRoot',
-      packageManagers['npm'],
+      getPackageManagerFactory('npm'),
       existsSyncMock,
       execaMock as never
     );
@@ -33,7 +33,7 @@ void describe('InitializedEnsurer', () => {
     await npmInitializedEnsurer.ensureInitialized();
     assert.equal(execaMock.mock.callCount(), 1);
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments, [
-      packageManagers['npm'],
+      getPackageManagerFactory('npm'),
       ['init', '--yes'],
       { stdin: 'inherit', cwd: '/testProjectRoot' },
     ]);
@@ -46,7 +46,7 @@ void describe('InitializedEnsurer', () => {
     });
     const npmInitializedEnsurer = new NpmProjectInitializer(
       '/testProjectRoot',
-      packageManagers['npm'],
+      getPackageManagerFactory('npm'),
       existsSyncMock,
       execaMock as never
     );
@@ -61,7 +61,7 @@ void describe('InitializedEnsurer', () => {
     const execaMock = mock.fn();
     const npmInitializedEnsurer = new NpmProjectInitializer(
       '/testProjectRoot',
-      packageManagers['npm'],
+      getPackageManagerFactory('npm'),
       existsSyncMock,
       execaMock as never
     );
